+++ app/console/app.php
@@ -20,8 +20,12 @@ $app['module']->register([
 
 $app['module']->addLoader(new AutoLoader($app['autoloader']));
 $app['module']->addLoader(new ConfigLoader(require $path.'/app/system/config.php'));
-$app['module']->addLoader(new ConfigLoader(require $app['config.file']));
-$app['module']->load(['system', 'console']);
+
+if ($app['config.file']) {
+    $app['module']->addLoader(new ConfigLoader(require $app['config.file']));
+    $app['module']->load('system');
+}
+$app['module']->load('console');
 
 $console = new Console($app, 'Pagekit');
 $console->run();