+++ app/installer/install.php
@@ -3,7 +3,7 @@
 use Pagekit\Application as App;
 
 App::config()->set('system/site', App::config('system/site')->merge([
-    'frontpage' => 1, 'view' => ['logo' => 'storage\/pagekit-logo.svg']
+    'frontpage' => 1, 'view' => ['logo' => 'storage/pagekit-logo.svg']
 ]));
 
 App::db()->insert('@system_config', ['name' => 'theme-one', 'value' => '{"logo_contrast":"storage\\/pagekit-logo-contrast.svg","_menus":{"main":"main","offcanvas":"main"},"_positions":{"hero":[1],"footer":[2]},"_widgets":{"1":{"title_hide":true,"title_size":"uk-panel-title","alignment":true,"html_class":"","panel":""},"2":{"title_hide":true,"title_size":"uk-panel-title","alignment":"","html_class":"","panel":""}},"_nodes":{"1":{"title_hide":true,"title_large":false,"alignment":true,"html_class":"","sidebar_first":false,"hero_image":"storage\\/home-hero.jpg","hero_viewport":true,"hero_contrast":true,"navbar_transparent":true}}}']);