+++ app/installer/requirements.php
@@ -434,6 +434,13 @@ class PagekitRequirements extends RequirementCollection
             'Install and enable the <strong>PDO</strong> extension.'
         );
 
+        if (version_compare($installedPhpVersion, '5.6', '>=') && version_compare($installedPhpVersion, '7.0.0', '<')) {
+            $this->addRequirement(!(ini_get('display_startup_errors') && ini_get('always_populate_raw_post_data') !== "-1"),
+                '\'display_startup_errors\' is enabled and \'always_populate_raw_post_data\' is not disabled',
+                'Disable startup errors or set \'always_populate_raw_post_data\' to \'-1\' in php.ini.'
+            );
+        }
+
         if (class_exists('PDO')) {
             $drivers = PDO::getAvailableDrivers();
             $this->addRequirement(