+++ app/system/modules/user/widgets/login.php
@@ -20,7 +20,7 @@ return [
     'render' => function ($widget) use ($app) {
 
         $user              = $app['user'];
-        $redirect          = $widget->get('redirect_login') ?: $app['url']->current(true);
+        $redirect          = $widget->get($user->isAuthenticated() ? 'redirect_logout' : 'redirect_login') ?: $app['url']->current(true);
         $last_username     = $app['session']->get(Auth::LAST_USERNAME);
         $remember_me_param = RememberMe::REMEMBER_ME_PARAM;